changed all of index to settings page and removed position label in image picker, also changed all imports to be from settings instead of index
